importer working quite well, implemented some security

+++ src/NetBS/CoreBundle/Controller/ActionController.php
@@ -2,6 +2,7 @@
 
 namespace NetBS\CoreBundle\Controller;
 
+use NetBS\SecureBundle\Voter\CRUD;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -22,6 +23,9 @@ class ActionController extends Controller
         $item       = $em->find($itemClass, $itemId);
 
         if($item) {
+
+            if(!$this->isGranted(CRUD::DELETE, $item))
+                throw $this->createAccessDeniedException();
 
             $em->remove($item);
             $em->flush();
