+++ src/NetBS/FichierBundle/Controller/TelephoneController.php
@@ -5,6 +5,7 @@ namespace NetBS\FichierBundle\Controller;
 use NetBS\CoreBundle\Utils\Modal;
 use NetBS\FichierBundle\Entity\Telephone;
 use NetBS\FichierBundle\Form\Contact\TelephoneType;
+use NetBS\SecureBundle\Voter\CRUD;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Symfony\Component\HttpFoundation\Request;
@@ -30,6 +31,9 @@ class TelephoneController extends Controller
         $owner  = $em->getRepository(base64_decode($ownerType))->find($ownerId);
         $tel    = $em->getRepository($class)->find($telephoneId);
 
+        if(!$this->isGranted(CRUD::DELETE, $tel))
+            throw $this->createAccessDeniedException("Suppression du numéro de téléphone refusée");
+
         $owner->removeTelephone($tel);
         $em->remove($tel);
         $em->flush();
@@ -53,6 +57,10 @@ class TelephoneController extends Controller
 
             $em     = $this->get('doctrine.orm.entity_manager');
             $holder = $em->getRepository(base64_decode($ownerType))->find($ownerId);
+
+            if(!$this->isGranted(CRUD::UPDATE, $holder))
+                throw $this->createAccessDeniedException("Accès refusé");
+
             $holder->addTelephone($form->getData());
 
             $em->persist($holder);
